requirements.txt
clearbit==0.1.7
celery==4.2
django==1.11
django-configurations==2.1
djangorestframework==3.6.0
djangorestframework-jwt==1.11.0
django-filter==2.0.0
gunicorn==19.9.0  
markdown==2.6.11
mixer==6.0.1
mock==2.0.0
psycopg2==2.7.5
redis==2.10.6
